CRUD and User Repository

+++ employee_management/src/main/java/com/ems/employee_management/service/EmployeeService.java
@@ -1,12 +1,11 @@
 package com.ems.employee_management.service;
-
+import com.ems.employee_management.entity.Employee;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.ems.employee_management.entity.Employee;
 import com.ems.employee_management.repository.EmployeeRepository;
 
 @Service
@@ -35,28 +34,41 @@ public class EmployeeService {
 				.orElseThrow(() -> new RuntimeException("Employee not found"));
 	}
 	
-	//updateEmployee
+	
+	
+	//update employee
 	public Employee updateEmployee(Long id, Employee updatedEmp) {
 		
-		Employee existing = employeeRepository.findById(id).orElseThrow(()-> new RuntimeException("Employee not found"));
+		Employee existing = employeeRepository.findById(id).orElseThrow(()-> new RuntimeException("Employee Not Found!!!"));
+		//not used existsById() as we require the employee object not just the boolean value provided by this method abt the existence of the object.
 		
-		    existing.setEmployeeId(updatedEmp.getEmployeeId());
-		    existing.setName(updatedEmp.getName());
-		    existing.setEmail(updatedEmp.getEmail());
-		    existing.setPhone(updatedEmp.getPhone());
-		    existing.setDepartment(updatedEmp.getDepartment());
-		    existing.setDesignation(updatedEmp.getDesignation());
-		    existing.setSalary(updatedEmp.getSalary());
-		    existing.setDateOfJoining(updatedEmp.getDateOfJoining());
+		existing.setName(updatedEmp.getName());
+		existing.setEmail(updatedEmp.getEmail());
+		existing.setDateOfJoining(updatedEmp.getDateOfJoining());
+		existing.setDepartment(updatedEmp.getDepartment());
+		existing.setDesignation(updatedEmp.getDesignation());
+		existing.setPhone(updatedEmp.getPhone());
+		existing.setSalary(updatedEmp.getSalary());
 		
-		return employeeRepository.save(existing);
-		
+		employeeRepository.save(existing);
+		return existing;
 	}
 	
 	//deleteEmployee 
-	public String deleteEmployee(Long id) {
+	public void deleteEmployee(Long id) {
 		
-		employeeRepository.deleteById(id);
-		return "Employee deleted successfully";
+//		if(employeeRepository.findById(id) != null) {
+//			employeeRepository.deleteById(id);
+//		}else {
+//			return "No such Employee found in the Records!";
+//		}
+//		return "Employee deleted successfully";
+		
+		if(!employeeRepository.existsById(id)) {
+			throw new RuntimeException("Employee not Found!");
+		}
+			employeeRepository.deleteById(id);
 	}
+
+	
 }
